refactor(desktop): index.ts から git/lsp/security/claudeHooks モジュールを分離
decision(module-structure): git/ と lsp/ はバレル化（index.ts re-export）、security と claudeHooks はフラットファイル
  — git/ は内部に status→worktree の依存があり、バレルで隠蔽する意味がある
  — security/claudeHooks は単一ファイルで完結するためディレクトリ不要
rejected(module-structure): 全モジュールを feature ディレクトリ化する案 — pty/fileServer/socketServer は単一ファイルで十分、ディレクトリにしても層が増えるだけ
rejected(watcher-extraction): watcher 群の切り出し — windowSwitchGen/windowLspClients 等のグローバル Map に密結合しており、コールバック注入の設計変更が必要。今回はスコープ外
constraint(claudeHooks): hookCommand 内で $ORKIS_SOCKET_PATH をシェル環境変数として参照しているため、関数引数として socketPath を渡す必要がない
