+++ apps/desktop/src/git/index.ts
@@ -1,0 +1,12 @@
+export { filterIgnored, getGitStatus, countChanges } from "./status";
+export {
+  WORKTREE_DIR,
+  generateWorktreeId,
+  addWorktree,
+  assertWorktreePath,
+  removeWorktree,
+  getWorktreeList,
+  attachChangeCounts,
+} from "./worktree";
+export { assertBranchName, getBranchList, deleteBranch } from "./branch";
+export { parseOwnerRepo } from "./utils";
