+++ apps/desktop/src/git/utils.ts
