+++ apps/desktop/src/lsp/index.ts
@@ -1,0 +1,4 @@
+export { createLspClient } from "./client";
+export type { LspClient } from "./client";
+export { resolveTsgoPath, resolveVueLspPaths } from "./resolve";
+export type { VueLspPaths } from "./resolve";
